SpMenu結合
